Move configuration initialization to structs
This moves the initialization of the configuration structs into the
individual structs for each configuration category.
