+++ src/config/key_bindings.rs
@@ -1,3 +1,6 @@
+use crate::config::utils::get_input;
+use git2::Config;
+
 #[derive(Clone, Debug)]
 pub(crate) struct KeyBindings {
 	pub(crate) abort: String,
@@ -26,4 +29,37 @@ pub(crate) struct KeyBindings {
 	pub(crate) rebase: String,
 	pub(crate) show_commit: String,
 	pub(crate) toggle_visual_mode: String,
+}
+
+impl KeyBindings {
+	pub(super) fn new(git_config: &Config) -> Result<Self, String> {
+		Ok(Self {
+			abort: get_input(&git_config, "interactive-rebase-tool.inputAbort", "q")?,
+			action_break: get_input(&git_config, "interactive-rebase-tool.inputActionBreak", "b")?,
+			action_drop: get_input(&git_config, "interactive-rebase-tool.inputActionDrop", "d")?,
+			action_edit: get_input(&git_config, "interactive-rebase-tool.inputActionEdit", "e")?,
+			action_fixup: get_input(&git_config, "interactive-rebase-tool.inputActionFixup", "f")?,
+			action_pick: get_input(&git_config, "interactive-rebase-tool.inputActionPick", "p")?,
+			action_reword: get_input(&git_config, "interactive-rebase-tool.inputActionReword", "r")?,
+			action_squash: get_input(&git_config, "interactive-rebase-tool.inputActionSquash", "s")?,
+			confirm_no: get_input(&git_config, "interactive-rebase-tool.inputConfirmNo", "n")?,
+			confirm_yes: get_input(&git_config, "interactive-rebase-tool.inputConfirmYes", "y")?,
+			edit: get_input(&git_config, "interactive-rebase-tool.inputEdit", "E")?,
+			force_abort: get_input(&git_config, "interactive-rebase-tool.inputForceAbort", "Q")?,
+			force_rebase: get_input(&git_config, "interactive-rebase-tool.inputForceRebase", "W")?,
+			help: get_input(&git_config, "interactive-rebase-tool.inputHelp", "?")?,
+			move_down: get_input(&git_config, "interactive-rebase-tool.inputMoveDown", "Down")?,
+			move_left: get_input(&git_config, "interactive-rebase-tool.inputMoveLeft", "Left")?,
+			move_right: get_input(&git_config, "interactive-rebase-tool.inputMoveRight", "Right")?,
+			move_up_step: get_input(&git_config, "interactive-rebase-tool.inputMoveStepUp", "PageUp")?,
+			move_down_step: get_input(&git_config, "interactive-rebase-tool.inputMoveStepDown", "PageDown")?,
+			move_selection_down: get_input(&git_config, "interactive-rebase-tool.inputMoveSelectionDown", "j")?,
+			move_selection_up: get_input(&git_config, "interactive-rebase-tool.inputMoveSelectionUp", "k")?,
+			move_up: get_input(&git_config, "interactive-rebase-tool.inputMoveUp", "Up")?,
+			open_in_external_editor: get_input(&git_config, "interactive-rebase-tool.inputOpenInExternalEditor", "!")?,
+			rebase: get_input(&git_config, "interactive-rebase-tool.inputRebase", "w")?,
+			show_commit: get_input(&git_config, "interactive-rebase-tool.inputShowCommit", "c")?,
+			toggle_visual_mode: get_input(&git_config, "interactive-rebase-tool.inputToggleVisualMode", "v")?,
+		})
+	}
 }
